requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
Jinja2==3.1.4
gunicorn==22.0.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
Flask-Limiter==3.9.0
